backend/requirements.txt
Flask==2.0.1
flask_cors==3.0.10
numpy==1.21.0
pandas==1.3.0
Pillow==8.3.1
Flask-Session
scipy
matplotlib
